# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: cfdmgrd API (cloudflared 多实例管理器)
  description: |
    无头 cloudflared 多实例管理器（headless cloudflared manager）。守护进程 `cfdmgrd` 把
    每个 frps 实例放进**独立的子进程**（`frps-worker`，re-exec 自身），通过
    HTTP/JSON 暴露 CRUD、生命周期、运行时监控、历史流量、告警、日志、导入导出、
    系统监控等管理面。

    关键架构事实（落到 API 字段命名、状态码、错误形态都需要先看一眼）：

    - **子进程模型**：每个运行中的 frps 都是父进程 re-exec 出来的独立子进程
      （`frps-worker`），原因是上游 `mem.StatsCollector` 是进程级全局单例，同进程
      跑多个 frps 会把所有实例流量混在一起，无法按实例分离。
    - **webServer 强制 loopback**：worker 启动前父进程预分配一个 `127.0.0.1:N`
      端口，强制覆盖 `webServer.addr/port/user/password`，账密随机；父进程通过
      该 loopback 反向读取 frps 原生 `/api/serverinfo`、`/api/proxy/{type}`、
      `/api/clients` 等运行时指标。**用户配置里的 webServer 项会被忽略**。
    - **reload = 重启**：frps 服务端参数（bindPort/vhost*/auth/tls 等）没有
      in-place 热重载语义，本守护进程对 `POST .../reload` 的实现就是先
      `stop()` 再 `start()`。
    - **每实例独立日志**：`<CFDM_DATA_DIR>/logs/<id>.log`，worker 的
      stdout/stderr 全量落盘。
    - **配置数据模型**：业务层（请求/响应里的 `config` 字段）= `ServerConfigV1`，
      内嵌上游 `v1.ServerConfig`，**全 camelCase**（bindPort/vhostHTTPPort/auth/
      transport/tls/...）。
    - **快照/事件/告警/流量**：`Snapshot`、`/runtime/*`、`/metrics/*`、`/alerts/*`、
      WebSocket `/events` 帧统一走 **snake_case**（id/inst_id/started_at/...）。
    - **`/runtime/*` 是 frps 原生 JSON 透传**：经 worker loopback 代理 frps 原生
      `/api/serverinfo`、`/api/proxy/{type}`、`/api/clients` 后**原样回写**，
      字段形态为 frps 原生 **camelCase**（不是本项目的 Snapshot snake_case 风格），
      参考 frps 文档 / `fatedier/frp` 源码。

...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/runtime/{id}/proxies:
    parameters:
      - $ref: '#/components/parameters/IdInPath'
    get:
      tags: [runtime]
      summary: 跨 8 个 proxy type 聚合 `/api/proxy/{type}` 后扁平化
      description: |
        守护进程顺次访问 frps `/api/proxy/{tcp|udp|http|https|stcp|sudp|xtcp|tcpmux}`，
        把每个类型返回的 `proxies` 数组拼成一个扁平列表。**元素形态为 frps 原生
        camelCase**（name/type/conf/...）。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [proxies]
                properties:
                  proxies:
                    type: array
                    items: { type: object, additionalProperties: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502': { $ref: '#/components/responses/UpstreamFailure' }
...
  /api/v1/runtime/{id}/clients:
    parameters:
      - $ref: '#/components/parameters/IdInPath'
    get:
      tags: [runtime]
      summary: 透传 frps `/api/clients`（活跃 frpc 客户端列表，camelCase 原生 JSON）
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502': { $ref: '#/components/responses/UpstreamFailure' }

  # ============================================================
  # 5. 历史流量（SQLite 时序）
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: '请求头 `Authorization: Bearer <CFDM_API_TOKEN>`'
  parameters:
    IdInPath:
      in: path
      name: id
      required: true
      schema: { type: string }
      description: "实例 ID（文件名去扩展名得到，禁用路径分隔符与 shell 特殊字符；长度 ≤ 64）"
  responses:
...
    Unauthorized:
      description: 缺失或无效的 Bearer Token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: 实例 / 规则不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: 状态机或资源冲突（如未运行 / 已存在）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UpstreamFailure:
      description: 经 worker loopback 访问 frps 失败 / 远程下载失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - validation_failed
            - internal_error
            - config_not_found
            - config_already_exists
            - invalid_state
            - proxy_not_found
            - proxy_already_exists
            - upstream_failure
        message: { type: string }
        details:
          type: object
          additionalProperties: true
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }

    # ----- 实例运行快照（snake_case） -----
...